aarch64: Fix 64-bit pointer related casts
BUG=None
BRANCH=none
TEST=Ran image in foundation model

Reviewed-on: https://chromium-review.googlesource.com/185273
Commit-Queue: Marcelo Póvoa <marcelogp@chromium.org>
Tested-by: Marcelo Póvoa <marcelogp@chromium.org>

+++ src/lib/selfboot.c
@@ -332,7 +332,7 @@ static int build_self_segment_list(
 			new->s_memsz = ntohl(segment->mem_len);
 			new->compression = ntohl(segment->compression);
 
-			new->s_srcaddr = (u32) ((unsigned char *)first_segment)
+			new->s_srcaddr = (uintptr_t) ((unsigned char *)first_segment)
 						+ ntohl(segment->offset);
 			new->s_filesz = ntohl(segment->len);
 			printk(BIOS_DEBUG, "  New segment dstaddr 0x%lx memsize 0x%lx srcaddr 0x%lx filesize 0x%lx\n",
@@ -526,7 +526,7 @@ void *selfload(struct lb_memory *mem, struct cbfs_payload *payload)
 
 	printk(BIOS_SPEW, "Loaded segments\n");
 
-	return (void *)entry;
+	return (void *)(uintptr_t)entry;
 
 out:
 	return NULL;
